% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekcelltype.r
\name{seekcelltype}
\alias{seekcelltype}
\title{Cell type annotation with Deepseek models}
\usage{
seekcelltype(
  input,
  tissuename = NULL,
  base_url = "https://api.deepseek.com",
  model = "deepseek-chat",
  DEEPSEEK_API_KEY = NULL,
  system_prompt = "You are an expert bioinformatician in single-cell RNA data analysis",
  seed = 24,
  topgenenumber = 20,
  ...
)
}
\arguments{
\item{input}{Either the differential gene table returned by Seurat FindAllMarkers() function, or a list of genes.}

\item{tissuename}{Optional input of tissue name.}

\item{base_url}{The url for Deepseek model. Default is 'https://api.deepseek.com'.}

\item{model}{The Deepseek model. Default is 'deepseek-chat'.}

\item{DEEPSEEK_API_KEY}{The Deepseek key. The default is NULL, which will resulting outputing the prompt itself. If an actual key is provided, then the output will be the celltype annotations from the GPT model specified by the user.}

\item{system_prompt}{Optional A system prompt to set the behavior of the assistant. Default is 'You are an expert bioinformatician in single-cell RNA data analysis'}

\item{seed}{Optional integer seed that Deepseek uses to try and make output more reproducible.}

\item{topgenenumber}{Number of top differential genes to be used if input is Seurat differential genes. Default is 20.}

\item{...}{Optional The parameters used in ellmer::chat_openai().}
}
\value{
A vector of cell types when the user provide Deepseek_key or the prompt itself when DEEPSEEK_API_KEY = NULL (default)
}
\description{
Annotate cell types by OpenAI GPT models in a Seurat pipeline or with a custom gene list. If used in a Seurat pipeline, Seurat FindAllMarkers() function needs to be run first and the differential gene table generated by Seurat will serve as the input. If the input is a custom list of genes, one cell type is identified for each element in the list.  Note: system environment should have variable DEEPSEEK_API_KEY = 'your_api_key' or ''. If '', then output the prompt itself. If an actual key is provided, then the output will be the celltype annotations from the Deepseek model specified by the user.
}
\details{
Cell type annotation with Deepseek models
}
\references{
Hou, W. and Ji, Z., 2023. Reference-free and cost-effective automated cell type annotation with GPT-4 in single-cell RNA-seq analysis. Nature Methods, 2024 March 25.
}
\author{
Wang Li \href{mailto:liwang@nibs.ac.cn}{liwang@nibs.ac.cn}
}
