% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seekcelltype.r
\name{seekcelltype_query}
\alias{seekcelltype_query}
\title{Cell type annotation summarization after multiple annotations with Deepseek models}
\usage{
seekcelltype_query(
  input,
  tissuename = NULL,
  base_url = "https://api.deepseek.com",
  anno_model = "deepseek-reasoner",
  query_model = "deepseek-reasoner",
  DEEPSEEK_API_KEY = NULL,
  seed = 24,
  topgenenumber_min = 20,
  topgenenumber_max = 30,
  Nquery = 6,
  system_prompt = "You are an expert bioinformatician in single-cell RNA data analysis",
  ...
)
}
\arguments{
\item{input}{Either the differential gene table returned by Seurat FindAllMarkers() function, or a list of genes.}

\item{tissuename}{input of tissue name.}

\item{base_url}{The url for Deepseek model. Default is 'https://api.deepseek.com'.}

\item{anno_model}{The Deepseek model for annotation. Default is 'deepseek-reasoner'. The optional models include 'deepseek-chat', 'deepseek-reasoner'.}

\item{query_model}{The Deepseek model for multiple query summarization. Default is 'deepseek-reasoner'. The optional models include 'deepseek-chat', 'deepseek-reasoner'.}

\item{DEEPSEEK_API_KEY}{The Deepseek key. The default is NULL, which will resulting outputing the prompt itself. If an actual key is provided, then the output will be the celltype annotations from the GPT model specified by the user.}

\item{seed}{Optional integer seed that Deepseek uses to try and make output more reproducible.}

\item{topgenenumber_min}{The minimum number of top differential genes to be used if input is Seurat differential genes. Default is 20.}

\item{topgenenumber_max}{The maximun number of top differential genes to be used if input is Seurat differential genes. Default is 30.}

\item{Nquery}{The number of rounds of multiple inquiries.}

\item{system_prompt}{Optional A system prompt to set the behavior of the assistant. Default is 'You are an expert bioinformatician in single-cell RNA data analysis'}

\item{...}{Optional The parameters used in ellmer::chat_openai().}
}
\value{
A vector of cell types with confidence score when the user provide Deepseek_key or the prompt itself when DEEPSEEK_API_KEY = NULL (default)
}
\description{
Annotate cell types by OpenAI GPT models in a Seurat pipeline or with a custom gene list. If used in a Seurat pipeline, Seurat FindAllMarkers() function needs to be run first and the differential gene table generated by Seurat will serve as the input. If the input is a custom list of genes, one cell type is identified for each element in the list.  Note: system environment should have variable DEEPSEEK_API_KEY = 'your_api_key' or ''. If '', then output the prompt itself. If an actual key is provided, then the output will be the celltype annotations from the Deepseek model specified by the user. The Cell type annotation summarization will be assessed after multiple inquiries. The output format is Cell Annotation-Confidence Score
}
\details{
Cell type annotation summarization after multiple annotations with Deepseek models
}
\references{
Hou, W. and Ji, Z., 2023. Reference-free and cost-effective automated cell type annotation with GPT-4 in single-cell RNA-seq analysis. Nature Methods, 2024 March 25.
}
\author{
Wang Li \href{mailto:liwang@nibs.ac.cn}{liwang@nibs.ac.cn}
}
